Move homepage nav card component to new dir
